backend/requirements.txt
fastapi
uvicorn
python-multipart
numpy
pillow
opencv-python-headless
scikit-image
scipy
yt-dlp
ffmpeg-python
